SYNCS.EXCH.64 64-bit src2 plumbing and GMMA fixes (#129)
* Enhance issue triage skill documentation for clarity and detail

* tracer: capture 64-bit src2 for SYNCS.EXCH.64 via generic aux field

Add auxRegVals[32] to the regular trace record as an opcode-defined
per-lane slot. SYNCS.EXCH.64 populates it with Rn+1 of its src2 pair;
the writer fuses srcRegVals[1] (lo) and auxRegVals (hi) into one 64-bit
column. Also enable reg-val tracing by default.


* Widen reg_val_t to uint64_t for SYNCS.EXCH.64 64-bit src2

Widen reg_val_t to std::array<uint64_t, WARP_SIZE> so the parser can
carry the full 64-bit src2 value for SYNCS.EXCH.64 (and future 64-bit
operand slots). Narrow explicitly at consumption sites that store into
32-bit operand fields (init.count, arrive.count/txCount,
complete_tx.txCount, wait.phase).


* Drop intermediate thread_counts copy in SYNCS.EXCH.64 reversal

Read trace.reg_src_vals[1][i] directly instead of copying into a
std::array<uint64_t, WARP_SIZE> local first.


* Capture mbarrier init_as_one flag from SYNCS.EXCH.64 trace

Per issue #123, the captured 64-bit register holding the mbarrier init
state encodes the pending-thread-count's top bit (bit 63) as a phase-
advance trigger. Extract that bit into syncs_operand.init_as_one for
the consumer in shader.cc to honor.

Also assert the lower 20 bits of the pending-thread-count field mirror
the expected-thread-count field, which is the only init encoding we
have observed. A failure here is not necessarily a bug — it just means
hardware produced a pattern we have not modeled and the decode below
needs to be extended.


* Handle large GMMA N sizes and gate SYNCS.EXCH.64 assert on lane mask

Compute latency/initiation interval as N/2 for GMMA N>=64 instead of
failing the lookup table. Skip the SYNCS.EXCH.64 pending-thread-count
sanity check for inactive lanes whose register values are undefined.


* traceDsm: write .traceg next to the input file by default

MODE_SIM now derives the output path from the input (.tracez -> .traceg)
instead of dumping to stdout, and rejects inputs without a .tracez
suffix. --annotate still goes to stdout.


* CI: drop redundant lastSuccess symlink loop and tighten failure skill

The post-correlate stage in main-sass.sh used to mv the workspace into
lastSuccess; that left $PWD empty so a follow-up loop symlinked the
files back. After the move was switched to rsync the workspace is no
longer emptied, but the symlink loop stayed and now fails with
"File exists" on the first item.

Also update ci-failure.md so the AI analyzer anchors on the literal
exit-1 error, knows about main-sass.sh / main-tracer.sh / h200.sh, and
checks whether sims actually succeeded before blaming simulator code.


---------

+++ gpu-simulator/trace-parser/trace_parser.h
@@ -64,7 +64,7 @@ struct tma_inst_memaddr_info_t {
                              const std::bitset<WARP_SIZE> &mask);
 };
 
-typedef std::array<uint32_t, WARP_SIZE> reg_val_t;
+typedef std::array<uint64_t, WARP_SIZE> reg_val_t;
 typedef enum {
   REG = 0,
   UREG,
